L3/requirements.txt
requests == 2.24.0
robotframework == 3.2.2
robotframework-archivelibrary == 0.4.0
robotframework-pabot == 1.8.1
robotframework-pythonlibcore == 2.1.0
robotframework-requests == 0.5.0
robotframework-seleniumlibrary == 4.5.0
selenium == 3.141.0
python-dotenv == 0.15.0